Rename package from token2.auth to token.auth2
