+++ auth-tokens-filter/src/test/resources/META-INF/services/io.dropwizard.logging.AppenderFactory
@@ -1,1 +1,1 @@
-com.palantir.tokens2.auth.http.MockAppenderFactory
+com.palantir.tokens.auth2.http.MockAppenderFactory
