+++ auth-tokens/src/main/java/com/palantir/tokens/auth2/Preconditions.java
@@ -28,7 +28,7 @@
  * the License.
  */
 
-package com.palantir.tokens2.auth;
+package com.palantir.tokens.auth2;
 
 /*
  * Derived from com.google.common.base.Preconditions
